Put all release tools and gh inside xaux, such that all other packages can use these directly. Also, added automatic version pinning.
